+++ src/record.hpp
@@ -1,7 +1,5 @@
 #pragma once
-
-#include <cstddef>
-#include <cstdint>
+#include <stdint.h>
 
 #define NDEF_RECORD_DECODE_ERROR_INVALID_LENGTH -2
 #define NDEF_RECORD_DECODE_ERROR_CHUNK_NOT_SUPPORTED -3
@@ -39,10 +37,10 @@ class NdefRecord
     int8_t set_type(uint8_t *type, uint8_t type_length);
     int8_t set_type(RTD type);
     int8_t set_id(uint8_t *id, uint8_t id_length);
-    int8_t set_payload(uint8_t *payload, size_t payload_length);
-    size_t get_encoded_size();
+    int8_t set_payload(uint8_t *payload, uint32_t payload_length);
+    uint32_t get_encoded_size();
     uint8_t *encode();
-    int8_t decode(uint8_t *data, size_t data_length);
+    int8_t decode(uint8_t *data, uint32_t data_length);
 
   private:
     TNF type_name_format;
@@ -51,5 +49,5 @@ class NdefRecord
     uint8_t *id;
     uint8_t id_length;
     uint8_t *payload;
-    size_t payload_length;
+    uint32_t payload_length;
 };
